+++ src/server/Server.Domain/Holdings/Holding.cs
@@ -10,9 +10,9 @@ namespace Server.Domain.Holdings;
 /// <param name="LastTransaction">The date of the latest contributing transaction.</param>
 public sealed record Holding(
     Symbol Symbol,
-    double Quantity,
-    double? AvgCost,
-    double TotalCost,
+    decimal Quantity,
+    decimal? AvgCost,
+    decimal TotalCost,
     int TransactionCount,
     string? FirstTransaction,
     string? LastTransaction);
